--- Merging r74475 into 'gcc/function.c': U    gcc/function.c
LLVM's optimizer does the inlining of "always_inline" functions. A function that
returns an aggregate value may not be a problem in this case. The GCC inliner
would normally inline these functions, and, if there are no more uses of said
function, remove it entirely so that it would never hit this warning.

For LLVM, check that the function is *not* marked "always_inline" before issuing
the warning.