Temporarily disable most use of region casts in RegionStoreManager, instead preferring to use the a region's actual type when creating symbols and using the QualType passed to Retrieve for implicit casting.  This preprocessor logic is temporary; eventually we will either remove region casts or keep them.
Temporarily toggle (via preprocessor directives) that SymbolicRegions
with heap storage are not undefined, but instead should be
symbolicated.  If we want to model that a SymbolicRegion is
uninitialized, we can explicitly model that by binding UndefinedVal to
that region.  It turns out that we want to treat most heap objects as
being defined, not the other way around.