People might want to do a "rotate" on a pointer. I don't know why either. This is a reduced testcase from real code:
void func(void *A) {
  unsigned int b = ((((unsigned long long) A) >> 5) | (((unsigned long long) A) << 27));
}

The problem is that GCC's front-end realizes that the size of A and the result
type are the same in 32-bit mode. So it strips the cast of A away. GCC also
realizes that this is a rotate operation. It marks it as so. However, LLVM
couldn't handle a rotate on of a pointer. It wants `A' to be of the same type as
`5' or `27' in this case. But that information was removed.

My "solution", such as it is, is to check for a pointer type in the EmitRotateOp
method. If it *is* a pointer, go ahead and convert the pointer to an integer
type that's the same as the constant. This way, we will get the rotate command
that we so desperately wish for.